[merge.c] fixed printing of N*w(M) on 32-bit machine
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1773 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/merge.c
@@ -215,10 +215,10 @@ main(int argc, char *argv[])
 #endif
 
     gzip_close(rep.outfile, outname);
-    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%lu\n",
+    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%"PRIu64"\n",
 	    mat.rem_nrows, mat.rem_ncols, mat.rem_nrows-mat.rem_ncols,
 	    mat.weight,
-	    ((unsigned long)mat.rem_nrows) * mat.weight);
+	    (uint64_t) mat.rem_nrows * (uint64_t) mat.weight);
 #if TEX
     fprintf(stderr, "\\end{verbatim}\n");
 #endif
